refactoring, moved ironmen calc function, moved UserSummary interface
